# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: RAG Evaluation Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /experiments:
    get:
      operationId: listExperiments
      tags: [experiments]
      summary: List all experiments
      responses:
        "200":
          description: List of experiments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Experiment"
    post:
      operationId: createExperiment
      tags: [experiments]
      summary: Create a new experiment configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateExperimentBody"
      responses:
        "201":
          description: Experiment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Experiment"
  /experiments/{id}:
    get:
      operationId: getExperiment
      tags: [experiments]
      summary: Get an experiment with its runs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Experiment detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentDetail"
    delete:
      operationId: deleteExperiment
      tags: [experiments]
      summary: Delete an experiment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /experiments/{id}/runs:
    post:
      operationId: createEvalRun
      tags: [eval-runs]
      summary: Trigger a new evaluation run for an experiment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEvalRunBody"
      responses:
        "201":
          description: Eval run created and started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvalRun"
  /eval-runs:
    get:
      operationId: listEvalRuns
      tags: [eval-runs]
      summary: List all evaluation runs
      parameters:
        - name: experimentId
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: List of eval runs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EvalRun"
  /eval-runs/{id}:
    get:
      operationId: getEvalRun
      tags: [eval-runs]
      summary: Get eval run with detailed per-question results
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Eval run detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvalRunDetail"
...
  /experiments/{id}/compare:
    get:
      operationId: compareExperimentRuns
      tags: [experiments]
      summary: Per-experiment, per-question comparison of all runs
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Per-question metrics across all runs of this experiment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExperimentRunComparison"
...
  /human-ratings:
    get:
      operationId: listHumanRatings
      tags: [human-ratings]
      summary: List ratings for an eval run from the current session
      parameters:
        - name: evalRunId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ratings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HumanRating"
    post:
      operationId: createHumanRating
      tags: [human-ratings]
      summary: Upsert a rating for a question in an eval run
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateHumanRatingBody"
      responses:
        "201":
          description: Rating saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HumanRating"
  /eval-runs/{id}/blind:
    get:
      operationId: getBlindEvalRun
      tags: [eval-runs]
      summary: Get per-question content for an eval run without metrics or config
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blind eval-run results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlindEvalRun"
...
components:
  schemas:
...
    Experiment:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        chunkSize:
          type: integer
        chunkOverlap:
          type: integer
        embeddingModel:
          type: string
        retrieverType:
          type: string
        topK:
          type: integer
        documentId:
          type: integer
        questionSetId:
          type: integer
        runCount:
          type: integer
        bestFaithfulness:
          type: number
          nullable: true
        bestContextRecall:
          type: number
          nullable: true
        isBlind:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - chunkSize
        - chunkOverlap
        - embeddingModel
        - retrieverType
        - topK
        - documentId
        - questionSetId
        - runCount
        - isBlind
        - createdAt
    ExperimentDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        chunkSize:
          type: integer
        chunkOverlap:
          type: integer
        embeddingModel:
          type: string
        retrieverType:
          type: string
        topK:
          type: integer
        documentId:
          type: integer
        questionSetId:
          type: integer
        isBlind:
          type: boolean
        runCount:
          type: integer
        bestFaithfulness:
          type: number
          nullable: true
        bestContextRecall:
          type: number
          nullable: true
        runs:
          type: array
          items:
            $ref: "#/components/schemas/EvalRun"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - chunkSize
        - chunkOverlap
        - embeddingModel
        - retrieverType
        - topK
        - documentId
        - questionSetId
        - isBlind
        - runCount
        - runs
        - createdAt
    CreateExperimentBody:
      type: object
      properties:
        name:
          type: string
        chunkSize:
          type: integer
        chunkOverlap:
          type: integer
        embeddingModel:
          type: string
        retrieverType:
          type: string
        topK:
          type: integer
        documentId:
          type: integer
        questionSetId:
          type: integer
      required:
        - name
        - chunkSize
        - chunkOverlap
        - embeddingModel
        - retrieverType
        - topK
        - documentId
        - questionSetId
    EvalRun:
      type: object
      properties:
        id:
          type: integer
        experimentId:
          type: integer
        status:
          type: string
          enum: [pending, running, completed, failed]
        avgFaithfulness:
          type: number
          nullable: true
        avgContextRecall:
          type: number
          nullable: true
        avgLatencyMs:
          type: number
          nullable: true
        totalQuestions:
          type: integer
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - experimentId
        - status
        - createdAt
    CreateEvalRunBody:
      type: object
      properties:
        questionSetId:
          type: integer
          nullable: true
      required: []
    EvalRunDetail:
      type: object
      properties:
        id:
          type: integer
        experimentId:
          type: integer
        status:
          type: string
          enum: [pending, running, completed, failed]
        avgFaithfulness:
          type: number
          nullable: true
        avgContextRecall:
          type: number
          nullable: true
        avgLatencyMs:
          type: number
          nullable: true
        results:
          type: array
          items:
            $ref: "#/components/schemas/EvalResult"
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - experimentId
        - status
        - results
        - createdAt
    EvalResult:
      type: object
      properties:
        id:
          type: integer
        evalRunId:
          type: integer
        questionId:
          type: integer
        questionText:
          type: string
        retrievedContext:
          type: string
          nullable: true
        generatedAnswer:
          type: string
          nullable: true
        faithfulness:
          type: number
          nullable: true
        contextRecall:
          type: number
          nullable: true
        latencyMs:
          type: number
          nullable: true
      required:
        - id
        - evalRunId
        - questionId
        - questionText
...
    ExperimentRunComparison:
      type: object
      properties:
        experimentId:
          type: integer
        experimentName:
          type: string
        runs:
          type: array
          items:
            type: object
            properties:
              runId:
                type: integer
              runNumber:
                type: integer
              status:
                type: string
              avgFaithfulness:
                type: number
                nullable: true
              avgContextRecall:
                type: number
                nullable: true
              avgLatencyMs:
                type: number
                nullable: true
              createdAt:
                type: string
                format: date-time
            required:
              - runId
              - runNumber
              - status
              - createdAt
        metricsByQuestion:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              questionText:
                type: string
              runMetrics:
                type: array
                items:
                  type: object
                  properties:
                    runId:
                      type: integer
                    faithfulness:
                      type: number
                      nullable: true
                    contextRecall:
                      type: number
                      nullable: true
                    latencyMs:
                      type: number
                      nullable: true
                  required:
                    - runId
            required:
              - questionId
              - questionText
              - runMetrics
      required:
        - experimentId
        - experimentName
        - runs
        - metricsByQuestion
...
    HumanRating:
      type: object
      properties:
        id:
          type: integer
        sessionId:
          type: string
        evalRunId:
          type: integer
        questionId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        preference:
          type: string
          enum: [A, B, tie]
          nullable: true
        arenaBattleId:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - sessionId
        - evalRunId
        - questionId
        - rating
        - createdAt
    CreateHumanRatingBody:
      type: object
      properties:
        evalRunId:
          type: integer
        questionId:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        preference:
          type: string
          enum: [A, B, tie]
        arenaBattleId:
          type: integer
      required:
        - evalRunId
        - questionId
        - rating
    BlindEvalRun:
      type: object
      properties:
        runId:
          type: integer
        questionResults:
          type: array
          items:
            type: object
            properties:
              questionId:
                type: integer
              questionText:
                type: string
              retrievedContext:
                type: string
                nullable: true
              generatedAnswer:
                type: string
                nullable: true
            required:
              - questionId
              - questionText
      required:
        - runId
        - questionResults
...